Draft Python scripts from Alex Vogel
I have now drafted the four Python functions that serve the purpose of performing SQL queries on the database and converting to JSONs that can be used for the html scripting. Since I do not have a database to look at I have used my own names and can modify once the database is set up. Please tell me if the functions need additional operations to be performed.